Added prompt for OpenAI API
