+++ config/settings/base.py
@@ -350,3 +350,5 @@ WAGTAILSEARCH_BACKENDS = {
 DEFAULT_AUTO_FIELD='django.db.models.AutoField'
 
 SITE_ID = 1
+
+OPENAI_API_KEY = os.getenv('OPENAI_API_SECRET_KEY')
