+++ ubyssey/urls.py
@@ -9,7 +9,9 @@ from wagtail import urls as wagtail_urls
 from wagtail.documents import urls as wagtaildocs_urls
 from wagtail.contrib.sitemaps.views import sitemap
 
-from ubyssey.views.main import ads_txt, redirect_blog_to_humour, publish_scheduled, get_image_urls
+from ubyssey.views.main import ads_txt, redirect_blog_to_humour, publish_scheduled
+from ubyssey.views.automate_adding_tags import get_image_urls
+
 from ubyssey.views.feed import FrontpageFeed, SectionFeed, AuthorFeed, TagFeed
 from ubyssey.views.advertise import AdvertiseTheme
 from ubyssey.views.tag import TagPage
